+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changes
 
+## 4.14.0
+
+### Application Changes
+
+- Added ID key/value pairs to the returned list of hosts and scorekeepers from `app.hosts.reports.appearances.retrieve_hosts()` and `app.scorekeepers.reports.appearances.retrieve_scorekeepers()`
+- Added "On This Day" omnibus report that includes the following for the current or specified day (via a separate route):
+  - Table of shows
+  - Tables of host, panelist and scorekeeper debuts
+  - Table of Not My Job guest scoring exceptions
+- Added `enable_on_this_day_report` application configuration flag that determines whether the On This Day report is available
+  - The report is enabled by default
+  - If the report is not enabled, the routes will redirect to the main index page
+- Corrected handling of the main `sitemap.xml` template to use `sitemapindex`
+  - Added `sitemaps.routes.main()` to cover main routes
+  - Added handling for the On This Day sitemap and conditional inclusion of entries based on the `enable_on_this_day_report` application configuration key
+  - Moved `app/sitemap/templates/sitemaps` to `app/templates/sitemaps` and update the Blueprint to match how sitemaps are handled in the Stats Page and Graphs Site applications
+- Refactored the "Not My Job Scoring Exceptions" to include Not My Job guests who have only appeared in Best Of shows
+  - Include show date, guest score, score exception flag and show notes directly in each exception dictionary
+  - Removed the no longer needed `app.guests.reports.retrieve_scoring_exceptions()` function
+  - Sorting of the exceptions is by show date, then by Not My Job guest name
+- Removed all straggling references and checks for `include_decimal_scores`
+- Removed the now redundant `app.scorekeepers.reports.appearances.retrieve_all_scorekeepers()`
+
+### Development Changes
+
+- Added Flask app to test fixtures in order to read application configuration settings in tests
+- Changed tests to use `response.text` rather than `response.data` when checking contents of the page to remove the need to use `b""`
+
 ## 4.13.0
 
 ### Application Changes